Decouple MapInstanceManager from window.farmOS.map #117
**Why?** Better support alternative bundling strategies
and scenarios where having mutable state in the global
`window.farmOS.map` is ugly/limiting.

These changes should allow farmOS-map to also be used
as follows;

```sh
npm install @farmos.org/farmos-map@2
```

```js
import MapInstanceManager from '@farmos.org/farmos-map/MapInstanceManager';

const maps = new MapInstanceManager();

const map = maps.create('my-map');
```

Some background can be found in the following issues;

* https://github.com/farmOS/farmOS-map/issues/113
* https://github.com/farmOS/farmOS-map/issues/68
