+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Remove redundant `defaults.behaviors` in favor of existing `farmOS.map.behaviors` mechanism for specifying the default behaviors.
 - Make instance behavior methods `async`.
 - Lazily load behaviors as separate Webpack chunks.
+- Decouple map instance creation/management from `window.farmOS.map`.
 
 ### Fixed
