+++ package.json
@@ -7,6 +7,7 @@
     "type": "git",
     "url": "https://github.com/farmOS/farmOS-map.git"
   },
+  "main": "src/index.js",
   "scripts": {
     "lint": "eslint src",
     "build": "npm run lint && webpack --config webpack.config.js --mode production",
