+++ src/index.js
@@ -1,0 +1,2 @@
+/* eslint-disable-next-line import/prefer-default-export */
+export { default as MapInstanceManager } from './MapInstanceManager';
